* All projects : set warning level to maximum   -> for MSVC set to level 4 (except for SQLite library, which must stay in level 3)
* Xpl : remove warnings
* Shared : remove warnings

+++ trunk/sources/external-libs/Webem/request_parser.cpp
@@ -315,7 +315,7 @@ boost::tribool request_parser::consume(request& req, char input)
 	  }
   case reading_content:
 	  req.content += input;
-	  if( req.content.length() == req.content_length )
+	  if( ((int)req.content.length()) == req.content_length )
 	  {
 		  return true;
 	  }
