+++ trunk/sources/shared/shared/Xpl/XplHubConnectedPeripheral.cpp
@@ -5,7 +5,7 @@
 #include "XplException.h"
 
 //m_socket(m_ioService), 
-CXplHubConnectedPeripheral::CXplHubConnectedPeripheral(boost::asio::io_service & io_service, std::string ip, int portNumber, int interval) 
+CXplHubConnectedPeripheral::CXplHubConnectedPeripheral(boost::asio::io_service & io_service, std::string ip, unsigned short portNumber, int interval) 
    :  m_portNumber(portNumber), m_interval(interval), m_ip(ip), m_socket(io_service), m_lastTimeSeen(boost::posix_time::second_clock::local_time())
 {
    m_socket.open(boost::asio::ip::udp::v4());
